% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chiSquareGoodnessOfFit.R
\name{chiSquareGoodnessOfFit}
\alias{chiSquareGoodnessOfFit}
\title{Chi-Square Goodness-of-Fit Test}
\usage{
chiSquareGoodnessOfFit(data, lambda, mu, bins = 10)
}
\arguments{
\item{data}{A numeric vector of observed data points.}

\item{lambda}{A numeric vector of lambda values for the Coxian distribution.}

\item{mu}{A numeric vector of mu values for the Coxian distribution.}

\item{bins}{Number of bins to use for the Chi-Square test. Default is 10.}
}
\value{
A list containing the Chi-Square statistic and p-value.
}
\description{
Computes the Chi-Square statistic for the goodness-of-fit of the Coxian distribution.
}
