% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_goodness_of_fit.R
\name{compute_goodness_of_fit}
\alias{compute_goodness_of_fit}
\title{Compute goodness-of-fit metrics (RMSE and KS test) for Coxian-PH model}
\usage{
compute_goodness_of_fit(fit, data)
}
\arguments{
\item{fit}{A fitted Coxian-PH model}

\item{data}{Observed survival times}
}
\value{
A list containing RMSE and Kolmogorov-Smirnov (KS) statistic
}
\description{
Compute goodness-of-fit metrics (RMSE and KS test) for Coxian-PH model
}
