% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_model_fit.R
\name{compute_model_fit}
\alias{compute_model_fit}
\title{Compute model fit statistics for a Coxian-PH model}
\usage{
compute_model_fit(fit, num_params, num_obs)
}
\arguments{
\item{fit}{A fitted Coxian-PH model object (output of fit_coxian)}

\item{num_params}{Number of estimated parameters}

\item{num_obs}{Number of observations}
}
\value{
A list with log-likelihood, AIC, and BIC values
}
\description{
Compute model fit statistics for a Coxian-PH model
}
