% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxianCdf.R
\name{coxianCdf}
\alias{coxianCdf}
\title{Coxian Cumulative Distribution Function}
\usage{
coxianCdf(t, lambda, mu)
}
\arguments{
\item{t}{A non-negative numeric vector.}

\item{lambda}{A numeric vector of the arrival rates.}

\item{mu}{A numeric vector of the service rates.}
}
\value{
The cumulative distribution function evaluated at t.
}
\description{
Coxian Cumulative Distribution Function
}
