% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxianPdf.R
\name{coxianPdf}
\alias{coxianPdf}
\title{Coxian Probability Density Function}
\usage{
coxianPdf(t, lambda, mu)
}
\arguments{
\item{t}{A non-negative numeric vector.}

\item{lambda}{A numeric vector of the arrival rates.}

\item{mu}{A numeric vector of the service rates.}
}
\value{
The probability density function evaluated at t.
}
\description{
Coxian Probability Density Function
}
