% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcoxian.R
\name{dcoxian}
\alias{dcoxian}
\title{Density of Coxian Phase Type Distribution}
\usage{
dcoxian(x, lambda, mu)
}
\arguments{
\item{x}{A numeric vector of observed data points.}

\item{lambda}{A numeric vector of lambda values.}

\item{mu}{A numeric vector of mu values.}
}
\value{
A numeric vector of density values for the provided data points.
}
\description{
Computes the density of the Coxian phase type distribution for a given vector of data points.
}
\examples{
lambda <- c(0.5, 0)
mu <- c(0.2, 0.4)
x_values <- seq(0, 10, by = 0.1)
density_values <- dcoxian(x_values, lambda, mu)

}
