% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_optimal_phases.R
\name{determine_optimal_phases}
\alias{determine_optimal_phases}
\title{Determine the optimal number of phases using Ross's formulation}
\usage{
determine_optimal_phases(data, max_phases = 5)
}
\arguments{
\item{data}{Observed survival times}

\item{max_phases}{Maximum number of phases to test}
}
\value{
A list containing the best-fitting model
}
\description{
Determine the optimal number of phases using Ross's formulation
}
