% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksGoodnessOfFit.R
\name{ksGoodnessOfFit}
\alias{ksGoodnessOfFit}
\title{Kolmogorov-Smirnov Goodness-of-Fit Test}
\usage{
ksGoodnessOfFit(data, lambda, mu)
}
\arguments{
\item{data}{A numeric vector of observed data points.}

\item{lambda}{A numeric vector of lambda values for the Coxian distribution.}

\item{mu}{A numeric vector of mu values for the Coxian distribution.}
}
\value{
A numeric value representing the KS statistic.
}
\description{
Computes the KS statistic for comparing the empirical CDF of the data with
the theoretical CDF of the fitted Coxian distribution.
}
