% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_function.R
\name{loss_function}
\alias{loss_function}
\title{Calculate Loss for Coxian Phase Type Distribution}
\usage{
loss_function(lambda, mu, data, method = "ssd")
}
\arguments{
\item{lambda}{A numeric vector of lambda values.}

\item{mu}{A numeric vector of mu values.}

\item{data}{A numeric vector of observed data points.}

\item{method}{A character string specifying the loss function to be used. Can take values
"nll" (negative log-likelihood) or "ssd" (sum of squared differences).}
}
\value{
The computed loss value.
}
\description{
Calculates the loss for the given lambda and mu parameters of a Coxian phase type distribution.
The loss can be computed as either the negative log-likelihood (NLL) or the sum of squared
differences (SSD).
}
\examples{
lambda <- c(0.5, 0)
mu <- c(0.2, 0.4)
data_sample <- rexp(100, rate = 0.5)
loss_value <- loss_function(lambda, mu, data_sample, method = "nll")

}
