% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{optimize}
\alias{optimize}
\title{Optimize Parameters for Coxian Phase Type Distribution}
\usage{
optimize(data, init_lambda, init_mu, objective = "ssd")
}
\arguments{
\item{data}{A numeric vector of observed data points.}

\item{init_lambda}{A numeric vector of initial values for lambda.}

\item{init_mu}{A numeric vector of initial values for mu.}

\item{objective}{A character string specifying the objective function to be minimized. Can
take values "nll" (negative log-likelihood) or "ssd" (sum of squared differences). Default
is "ssd".}
}
\value{
A list containing the optimized lambda and mu values.
}
\description{
This function optimizes the parameters lambda and mu of a Coxian phase type distribution
to fit the provided data. The objective function to be minimized
can be controlled through the `objective` parameter.
}
\examples{
data_sample <- rexp(100, rate = 0.5)
init_lambda <- c(0.5, 0)
init_mu <- c(0.2, 0.4)
result <- optimize(data_sample, init_lambda, init_mu, objective = "nll")

}
