% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcoxian.R
\name{pcoxian}
\alias{pcoxian}
\title{Cumulative Distribution Function of Coxian Phase Type Distribution}
\usage{
pcoxian(q, lambda, mu)
}
\arguments{
\item{q}{A numeric vector of observed data points.}

\item{lambda}{A numeric vector of lambda values.}

\item{mu}{A numeric vector of mu values.}
}
\value{
A numeric vector of CDF values for the provided data points.
}
\description{
Computes the cumulative distribution function (CDF) of the Coxian phase type distribution
for a given vector of data points.
}
\examples{
lambda <- c(0.5, 0)
mu <- c(0.2, 0.4)
q_values <- seq(0, 10, by = 0.1)
cdf_values <- pcoxian(q_values, lambda, mu)

}
