% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coxian.R
\name{plot_coxian}
\alias{plot_coxian}
\title{Plot the fitted Coxian Phase-Type distribution}
\usage{
plot_coxian(fit, max_t = 10)
}
\arguments{
\item{fit}{A fitted Coxian-PH model object from fit_coxian()}

\item{max_t}{Maximum time to display on the x-axis}
}
\value{
A ggplot object
}
\description{
Plot the fitted Coxian Phase-Type distribution
}
