% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_log_likelihood.R
\name{plot_log_likelihood}
\alias{plot_log_likelihood}
\title{Plot Log-Likelihood Progression}
\usage{
plot_log_likelihood(fit)
}
\arguments{
\item{fit}{A fitted Coxian-PH model object (output of fit_coxian)}
}
\value{
A ggplot object showing log-likelihood progression across iterations
}
\description{
Plot Log-Likelihood Progression
}
