% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_survival_fit.R
\name{plot_survival_fit}
\alias{plot_survival_fit}
\title{Plot Kaplan-Meier vs Coxian-PH survival curve}
\usage{
plot_survival_fit(fit, data)
}
\arguments{
\item{fit}{A fitted Coxian-PH model}

\item{data}{Survival time data used for fitting}
}
\description{
Plot Kaplan-Meier vs Coxian-PH survival curve
}
