% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHistVsTheoreticalPdf.R
\name{plotHistVsTheoreticalPdf}
\alias{plotHistVsTheoreticalPdf}
\title{Plot Histogram with Theoretical PDF Overlay}
\usage{
plotHistVsTheoreticalPdf(
  data,
  lambda,
  mu,
  breaks = 30,
  main = "Histogram vs Theoretical PDF",
  xlab = "Value",
  ylab = "Density",
  line_col = "blue",
  line_lwd = 2
)
}
\arguments{
\item{data}{A numeric vector of observed data points.}

\item{lambda}{A numeric vector of lambda values for the Coxian distribution.}

\item{mu}{A numeric vector of mu values for the Coxian distribution.}

\item{breaks}{The number of breaks for the histogram. Default is 30.}

\item{main}{The title of the plot. Default is "Histogram vs Theoretical PDF".}

\item{xlab}{The x-axis label. Default is "Value".}

\item{ylab}{The y-axis label. Default is "Density".}

\item{line_col}{The color of the PDF line. Default is "blue".}

\item{line_lwd}{The line width of the PDF line. Default is 2.}
}
\value{
A plot showing the histogram with the theoretical PDF overlay.
}
\description{
Plots a histogram of the observed data and overlays it with the theoretical
probability density function (PDF) of the fitted Coxian phase-type distribution.
}
\examples{
data_sample <- rexp(100, rate = 0.5)
lambda <- c(0.5, 0)
mu <- c(0.2, 0.4)
plotHistVsTheoreticalPdf(data_sample, lambda, mu)

}
