% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_coxian.R
\name{fit_coxian}
\alias{fit_coxian}
\title{Fit a Coxian Phase-Type model using Ross's formulation}
\usage{
fit_coxian(data, num_phases, max_iter = 500, tol = 1e-06)
}
\arguments{
\item{data}{Observed survival times}

\item{num_phases}{Number of phases}

\item{max_iter}{Maximum EM iterations}

\item{tol}{Convergence tolerance}
}
\value{
A list containing estimated parameters
}
\description{
Fit a Coxian Phase-Type model using Ross's formulation
}
