% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_survival_fit.R
\name{plot_survival_fit}
\alias{plot_survival_fit}
\title{Plot survival function using Ross's equation}
\usage{
plot_survival_fit(fit, data)
}
\arguments{
\item{fit}{Fitted Coxian-PH model}

\item{data}{Observed survival times}
}
\value{
A ggplot object
}
\description{
Plot survival function using Ross's equation
}
