% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_coxian.R
\name{simulate_coxian}
\alias{simulate_coxian}
\title{Simulate survival data using Ross's formulation for Coxian-PH distributions}
\usage{
simulate_coxian(n, lambda, mu)
}
\arguments{
\item{n}{Number of observations to generate}

\item{lambda}{Vector of transition rates}

\item{mu}{Vector of absorption rates}
}
\value{
A numeric vector of simulated survival times
}
\description{
Simulate survival data using Ross's formulation for Coxian-PH distributions
}
